+++ bitwise/__init__.py
@@ -2,9 +2,19 @@ from . import wire
 from . import gate
 from . import arithmetic
 from . import logic
+from . import processor
 from . import signal
 from . import state
 from . import storage
 from ._version import __version__
 
-__all__ = ["wire", "gate", "arithmetic", "logic", "signal", "state", "storage"]
+__all__ = [
+    "wire",
+    "gate",
+    "arithmetic",
+    "logic",
+    "processor",
+    "signal",
+    "state",
+    "storage"
+]