+++ bitwise/wire/BUS.py
@@ -25,11 +25,20 @@ class Bus4:
     """
     def __init__(
         self,
-        wire_1=Wire(),
-        wire_2=Wire(),
-        wire_3=Wire(),
-        wire_4=Wire()
+        wire_1=None,
+        wire_2=None,
+        wire_3=None,
+        wire_4=None
     ):
+        if not wire_1:
+            wire_1 = Wire()
+        if not wire_2:
+            wire_2 = Wire()
+        if not wire_3:
+            wire_3 = Wire()
+        if not wire_4:
+            wire_4 = Wire()
+
         self._wires = (wire_1, wire_2, wire_3, wire_4)
 
     @property
@@ -46,10 +55,10 @@ class Bus4:
             raise TypeError(
                 "Expected 4 arguments, received {0}.".format(len(values))
             )
-        self.wires[0].value = values[0]
-        self.wires[1].value = values[1]
-        self.wires[2].value = values[2]
-        self.wires[3].value = values[3]
+        self._wires[0].value = values[0]
+        self._wires[1].value = values[1]
+        self._wires[2].value = values[2]
+        self._wires[3].value = values[3]
 
     def __getitem__(self, key):
         if isinstance(key, slice):
@@ -63,6 +72,9 @@ class Bus4:
 
     def __len__(self):
         return len(self._wires)
+
+    def __str__(self):
+        return str(self.wire_values)
 
 
 class Bus8:
@@ -79,15 +91,32 @@ class Bus8:
     """
     def __init__(
         self,
-        wire_1=Wire(),
-        wire_2=Wire(),
-        wire_3=Wire(),
-        wire_4=Wire(),
-        wire_5=Wire(),
-        wire_6=Wire(),
-        wire_7=Wire(),
-        wire_8=Wire()
+        wire_1=None,
+        wire_2=None,
+        wire_3=None,
+        wire_4=None,
+        wire_5=None,
+        wire_6=None,
+        wire_7=None,
+        wire_8=None
      ):
+        if not wire_1:
+            wire_1 = Wire()
+        if not wire_2:
+            wire_2 = Wire()
+        if not wire_3:
+            wire_3 = Wire()
+        if not wire_4:
+            wire_4 = Wire()
+        if not wire_5:
+            wire_5 = Wire()
+        if not wire_6:
+            wire_6 = Wire()
+        if not wire_7:
+            wire_7 = Wire()
+        if not wire_8:
+            wire_8 = Wire()
+
         self._wires = (
             wire_1,
             wire_2,
@@ -113,14 +142,14 @@ class Bus8:
             raise TypeError(
                 "Expected 8 arguments, received {0}.".format(len(values))
             )
-        self.wires[0].value = values[0]
-        self.wires[1].value = values[1]
-        self.wires[2].value = values[2]
-        self.wires[3].value = values[3]
-        self.wires[4].value = values[4]
-        self.wires[5].value = values[5]
-        self.wires[6].value = values[6]
-        self.wires[7].value = values[7]
+        self._wires[0].value = values[0]
+        self._wires[1].value = values[1]
+        self._wires[2].value = values[2]
+        self._wires[3].value = values[3]
+        self._wires[4].value = values[4]
+        self._wires[5].value = values[5]
+        self._wires[6].value = values[6]
+        self._wires[7].value = values[7]
 
     def __getitem__(self, key):
         if isinstance(key, slice):
@@ -134,6 +163,9 @@ class Bus8:
 
     def __len__(self):
         return len(self._wires)
+
+    def __str__(self):
+        return str(self.wire_values)
 
 
 class Bus16:
@@ -150,23 +182,56 @@ class Bus16:
     """
     def __init__(
         self,
-        wire_1=Wire(),
-        wire_2=Wire(),
-        wire_3=Wire(),
-        wire_4=Wire(),
-        wire_5=Wire(),
-        wire_6=Wire(),
-        wire_7=Wire(),
-        wire_8=Wire(),
-        wire_9=Wire(),
-        wire_10=Wire(),
-        wire_11=Wire(),
-        wire_12=Wire(),
-        wire_13=Wire(),
-        wire_14=Wire(),
-        wire_15=Wire(),
-        wire_16=Wire()
+        wire_1=None,
+        wire_2=None,
+        wire_3=None,
+        wire_4=None,
+        wire_5=None,
+        wire_6=None,
+        wire_7=None,
+        wire_8=None,
+        wire_9=None,
+        wire_10=None,
+        wire_11=None,
+        wire_12=None,
+        wire_13=None,
+        wire_14=None,
+        wire_15=None,
+        wire_16=None
      ):
+        if not wire_1:
+            wire_1 = Wire()
+        if not wire_2:
+            wire_2 = Wire()
+        if not wire_3:
+            wire_3 = Wire()
+        if not wire_4:
+            wire_4 = Wire()
+        if not wire_5:
+            wire_5 = Wire()
+        if not wire_6:
+            wire_6 = Wire()
+        if not wire_7:
+            wire_7 = Wire()
+        if not wire_8:
+            wire_8 = Wire()
+        if not wire_9:
+            wire_9 = Wire()
+        if not wire_10:
+            wire_10 = Wire()
+        if not wire_11:
+            wire_11 = Wire()
+        if not wire_12:
+            wire_12 = Wire()
+        if not wire_13:
+            wire_13 = Wire()
+        if not wire_14:
+            wire_14 = Wire()
+        if not wire_15:
+            wire_15 = Wire()
+        if not wire_16:
+            wire_16 = Wire()
+
         self._wires = (
             wire_1,
             wire_2,
@@ -200,22 +265,22 @@ class Bus16:
             raise TypeError(
                 "Expected 16 arguments, received {0}.".format(len(values))
             )
-        self.wires[0].value = values[0]
-        self.wires[1].value = values[1]
-        self.wires[2].value = values[2]
-        self.wires[3].value = values[3]
-        self.wires[4].value = values[4]
-        self.wires[5].value = values[5]
-        self.wires[6].value = values[6]
-        self.wires[7].value = values[7]
-        self.wires[8].value = values[8]
-        self.wires[9].value = values[9]
-        self.wires[10].value = values[10]
-        self.wires[11].value = values[11]
-        self.wires[12].value = values[12]
-        self.wires[13].value = values[13]
-        self.wires[14].value = values[14]
-        self.wires[15].value = values[15]
+        self._wires[0].value = values[0]
+        self._wires[1].value = values[1]
+        self._wires[2].value = values[2]
+        self._wires[3].value = values[3]
+        self._wires[4].value = values[4]
+        self._wires[5].value = values[5]
+        self._wires[6].value = values[6]
+        self._wires[7].value = values[7]
+        self._wires[8].value = values[8]
+        self._wires[9].value = values[9]
+        self._wires[10].value = values[10]
+        self._wires[11].value = values[11]
+        self._wires[12].value = values[12]
+        self._wires[13].value = values[13]
+        self._wires[14].value = values[14]
+        self._wires[15].value = values[15]
 
     def __getitem__(self, key):
         if isinstance(key, slice):
@@ -230,6 +295,9 @@ class Bus16:
     def __len__(self):
         return len(self._wires)
 
+    def __str__(self):
+        return str(self.wire_values)
+
 
 class BusSevenSegmentDisplay:
     """Initialize a new seven-segment display bus.
@@ -243,14 +311,29 @@ class BusSevenSegmentDisplay:
     """
     def __init__(
         self,
-        wire_1=Wire(),
-        wire_2=Wire(),
-        wire_3=Wire(),
-        wire_4=Wire(),
-        wire_5=Wire(),
-        wire_6=Wire(),
-        wire_7=Wire()
+        wire_1=None,
+        wire_2=None,
+        wire_3=None,
+        wire_4=None,
+        wire_5=None,
+        wire_6=None,
+        wire_7=None
      ):
+        if not wire_1:
+            wire_1 = Wire()
+        if not wire_2:
+            wire_2 = Wire()
+        if not wire_3:
+            wire_3 = Wire()
+        if not wire_4:
+            wire_4 = Wire()
+        if not wire_5:
+            wire_5 = Wire()
+        if not wire_6:
+            wire_6 = Wire()
+        if not wire_7:
+            wire_7 = Wire()
+
         self._wires = (
             wire_1,
             wire_2,
@@ -275,14 +358,14 @@ class BusSevenSegmentDisplay:
             raise TypeError(
                 "Expected 7 arguments, received {0}.".format(len(values))
             )
-        self.wires[0].value = values[0]
-        self.wires[1].value = values[1]
-        self.wires[2].value = values[2]
-        self.wires[3].value = values[3]
-        self.wires[4].value = values[4]
-        self.wires[5].value = values[5]
-        self.wires[6].value = values[6]
-        self.wires[7].value = values[7]
+        self._wires[0].value = values[0]
+        self._wires[1].value = values[1]
+        self._wires[2].value = values[2]
+        self._wires[3].value = values[3]
+        self._wires[4].value = values[4]
+        self._wires[5].value = values[5]
+        self._wires[6].value = values[6]
+        self._wires[7].value = values[7]
 
     def __getitem__(self, key):
         if isinstance(key, slice):
@@ -296,3 +379,6 @@ class BusSevenSegmentDisplay:
 
     def __len__(self):
         return len(self._wires)
+
+    def __str__(self):
+        return str(self.wire_values)