+++ docs/changelog.rst
@@ -11,7 +11,7 @@ Unreleased
 
 Added
 -----
-* Arithmetic-logic unit to processor subpackage
+* Processor, arithmetic-logic unit, and 10-bit bus modules to processor subpackage
 * Bitwise logic operations to logic subpackage
     * Bitwise AND, NAND, NOR, NOT, OR, XNOR, and XOR operations for 4-, 8-, and 16- bit inputs
 * Multiplier classes to arithmetic subpackage